airflow/requirements.txt
# Provider-compatible pins for Airflow 2.9.2
pandas==2.1.4          # google provider requires <2.2
requests==2.31.0       # docker provider requires <2.32
PyYAML==6.0.2
Jinja2==3.1.4
atlassian-python-api==3.41.16
tenacity==9.0.0
gitpython==3.1.43
fastjsonschema==2.20.0

